Fixes for GCC 11

+++ src/ps2cnf.c
@@ -15,7 +15,7 @@
 
 static const char *CNFGetToken(const char *cnf, const char *key)
 {
-    for (; isspace(*cnf); cnf++) {
+    for (; isspace((int)*cnf); cnf++) {
     }
 
     for (; *key != '\0'; key++, cnf++) {
@@ -45,7 +45,7 @@ static const char *CNFGetKey(const char *line, char *key)
     int i;
 
     //Skip leading whitespace
-    for (; isspace(*line); line++) {
+    for (; isspace((int)*line); line++) {
     }
 
     if (*line == '\0') { //Unexpected end of file
@@ -53,11 +53,11 @@ static const char *CNFGetKey(const char *line, char *key)
     }
 
     for (i = 0; i < CNF_PATH_LEN_MAX && *line != '\0'; i++) {
-        if (isgraph(*line)) {
+        if (isgraph((int)*line)) {
             *key = *line;
             line++;
             key++;
-        } else if (isspace(*line)) {
+        } else if (isspace((int)*line)) {
             *key = '\0';
             break;
         } else if (*line == '\0') { //Unexpected end of file. This check exists, along with the other similar check above.
